requirements.txt
asgiref==3.6.0
attrs==22.2.0
coverage==7.0.3
Django==4.1.5
djangorestframework==3.14.0
exceptiongroup==1.1.0
iniconfig==1.1.1
packaging==22.0
parameterized==0.8.1
parametrized==0.1
pluggy==1.0.0
pytest==7.2.0
pytest-cov==4.0.0
pytest-django==4.5.2
pytz==2022.7
sqlparse==0.4.3
tomli==2.0.1
